chore: move common get token functions to helpers
